feat: use typescript in test cases
