+++ cypress/e2e/03-developer-user-permissions-setup.cy.ts
