+++ cypress/e2e/04-developer-user-signup.cy.ts
