+++ cypress/e2e/06-add-devices.cy.ts
