+++ cypress/e2e/07-add-history-meter-read.cy.ts
