--- cypress/e2e/13-device-bulk-upload.cy.js
@@ -1,10 +1,0 @@
-describe('Bulk Upload Test', () => {
-    beforeEach(() => {
-        cy.wait(1000);
-      });
-    it('should upload a file', () => {
-      cy.developerUserLogin().wait(60000);
-      cy.bulkUpload('device-bulk-upload-template'); 
-    });
-  });
-
