+++ cypress/e2e/13-device-bulk-upload.cy.ts
@@ -1,0 +1,9 @@
+describe('Bulk Upload Test', () => {
+  beforeEach(() => {
+    cy.wait(1000);
+  });
+  it('should upload a file', () => {
+    cy.developerUserLogin().wait(60000);
+    cy.bulkUpload('device-bulk-upload-template');
+  });
+});
