--- cypress/fixtures/account-settings.js
@@ -1,47 +1,0 @@
-[
-  {
-    action: 'click',
-    selector: '[test-id="account-setting-header"]'
-  },
-  {
-    action: 'click',
-    selector: '[test-id="profile"]'
-  },
-  {
-    action: 'type',
-    selector: '[test-id="first-name"]',
-    value: 'test'
-  },
-  {
-    action: 'type',
-    selector: '[test-id="last-name"]',
-    value: 'update'
-  },
-  {
-    action: 'type',
-    selector: '[test-id="email"]',
-    value: 'testupdate@drc.org'
-  },
-  {
-    action: 'click',
-    selector: '[test-id="update-profile"]'
-  },
-  {
-    action: 'click',
-    selector: '.mdc-tab__text-label:contains("Security Setting")'
-  },  
-  {
-    action: 'type',
-    selector: '[test-id="new-password"]',
-    value: '1234@Drec'
-  },
-  {
-    action: 'type',
-    selector: '[test-id="confirm-password"]',
-    value: '1234@Drec'
-  },
-  {
-    action: 'click',
-    selector: '[test-id="reset-password-form"]'
-  }
-]
