+++ cypress/fixtures/add-aggregate-meter-read.json
@@ -1,0 +1,45 @@
+[
+    {
+      "action": "click",
+      "selector": "[test-id='meter-read']"
+    },
+  
+    {
+      "action": "click",
+      "selector": "[test-id='add-meter-read']"
+    },
+    {
+      "action": "selected",
+      "selector": "[test-id='search-external-id']",
+      "option": "[external-id-test-id]"
+    },
+    {
+      "action": "select-timezone",
+      "selector": "[test-id='enter-time-zone']",
+      "option": "[timezone-test-id]"
+    },
+    {
+      "action": "select",
+      "selector": "mat-select[formControlName='type']",
+      "option": "[read-type-test-id]"
+    },
+    {
+      "action": "select",
+      "selector": "mat-select[formControlName='unit']",
+      "option": "[unit-test-id]"
+    },
+    {
+      "action": "type",
+      "selector": "[test-id='read-value']",
+      "value": "100"
+    },
+    {
+      "action": "date-picker",
+      "selector": "[test-id='choose-end-date']"
+    },
+    {
+      "action": "click",
+      "selector": "[test-id='submit-read']"
+    }
+  ]
+
