--- cypress/fixtures/add-device.js
@@ -1,70 +1,0 @@
-[
-  {
-    action: 'click',
-    selector: '[test-id="devices"]',
-  },
-
-  {
-    action: 'click',
-    selector: '[test-id="add-devices"]',
-  },
-  {
-    action: 'type',
-    selector: '[test-id="project-name"]',
-    value: 'test-device',
-  },
-  {
-    action: 'type',
-    selector: '[test-id="external-id"]',
-    value: 'device test',
-  },
-  {
-    action: 'select',
-    selector: '[test-id="select-country"]',
-    option: '[test-id="country-options"]',
-  },
-  {
-    action: 'write',
-    selector: '[test-id="commissioning-date"]',
-    value: '10/31/2024, 7:17:19 PM',
-  },
-  {
-    action: 'type',
-    selector: '[test-id="capacity"]',
-    value: '100',
-  },
-  {
-    action: 'type',
-    selector: '[test-id="address"]',
-    value: 'just testing',
-  },
-  {
-    action: 'type',
-    selector: '[test-id="latitude"]',
-    value: '21.00876',
-  },
-  {
-    action: 'type',
-    selector: '[test-id="longitude"]',
-    value: '31.00876',
-  },
-  {
-    action: 'select',
-    selector: '[test-id="fuel-code"]',
-    option: '[test-id="fuel-option"]',
-  },
-  {
-    action: 'select',
-    selector: '[test-id="device-type"]',
-    option: '[test-id="type-code-option"]',
-  },
-  {
-    action: 'select',
-    selector: '[test-id="SDGBenefits"]',
-    option: '[test-id="SDGBenefits-option"]',
-  },
-  {
-    action: 'submit',
-    selector: '[test-id="submit-device"]',
-  },
-]
