+++ cypress/fixtures/add-history-meter-read.json
@@ -1,0 +1,50 @@
+[
+  {
+    "action": "click",
+    "selector": "[test-id='meter-read']"
+  },
+
+  {
+    "action": "click",
+    "selector": "[test-id='add-meter-read']"
+  },
+  {
+    "action": "selected",
+    "selector": "[test-id='search-external-id']",
+    "option": "[external-id-test-id]"
+  },
+  {
+    "action": "select-timezone",
+    "selector": "[test-id='enter-time-zone']",
+    "option": "[timezone-test-id]"
+  },
+  {
+    "action": "select",
+    "selector": "mat-select[formControlName='type']",
+    "option": "[read-type-test-id]"
+  },
+  {
+    "action": "select",
+    "selector": "mat-select[formControlName='unit']",
+    "option": "[unit-test-id]"
+  },
+  {
+    "action": "type",
+    "selector": "[test-id='read-value']",
+    "value": "100"
+  },
+  {
+    "action": "type",
+    "selector": "[test-id='choose-start-date']",
+    "value": "12/10/2024, 12:43:54 PM"
+  },
+  {
+    "action": "type",
+    "selector": "[test-id='choose-end-date']",
+    "value": "1/6/2025, 12:44:55 PM"
+  },
+  {
+    "action": "click",
+    "selector": "[test-id='submit-read']"
+  }
+]
