--- cypress/fixtures/admin-login.js
@@ -1,20 +1,0 @@
-[
-  {
-    action: 'type',
-    selector: '[test-id="login-username"]',
-    index: 0,
-  },
-  {
-    action: 'type',
-    selector: '[test-id="login-password"]',
-    index: 1,
-  },
-  {
-    action: 'click',
-    selector: '[test-id="hide-password"]',
-  },
-  {
-    action: 'click',
-    selector: '[test-id="login-submit"]'
-  }
-]
