+++ cypress/fixtures/admin-login.json
@@ -1,0 +1,20 @@
+[
+  {
+    "action": "type",
+    "selector": "[test-id='login-username']",
+    "index": 0
+  },
+  {
+    "action": "type",
+    "selector": "[test-id='login-password']",
+    "index": 1
+  },
+  {
+    "action": "click",
+    "selector": "[test-id='hide-password']"
+  },
+  {
+    "action": "click",
+    "selector": "[test-id='login-submit']"
+  }
+]
