+++ cypress/fixtures/bulk-upload.json
@@ -1,0 +1,22 @@
+[
+  {
+    "action": "click",
+    "selector": "[test-id='devices']"
+  },
+  {
+    "action": "click",
+    "selector": "[test-id='bulk-upload']"
+  },
+  {
+    "action": "click",
+    "selector": "[test-id='browse']"
+  },
+  {
+    "action": "upload",
+    "selector": "[test-id='file-name']"
+  },
+  {
+    "action": "click",
+    "selector": "[test-id='upload']"
+  }
+]
