+++ cypress/fixtures/buyer-user-permissions-setup.json
@@ -1,0 +1,86 @@
+[
+    {
+      "action": "click",
+      "selector": "[test-id='permission']"
+    },
+    {
+      "action": "click",
+      "selector": "[test-id='user-role-permission']"
+    },
+    {
+      "action": "click",
+      "selector": "[aria-label='add-user-role-permissions-test-id']"
+    },
+    {
+      "action": "select",
+      "selector": "[test-id='add-user-role']",
+      "option": "[test-id='user-role-option']"
+    },
+    {
+      "action": "check",
+      "selector": "[test-id='select-module']",
+      "index": 4
+    },
+    {
+      "action": "check-multiple",
+      "selector": "[test-id='select-permission']",
+      "contains": "BUYER_RESERVATION_MANAGEMENT_CRUDL"
+    },
+    {
+      "action": "check",
+      "selector": "[test-id='select-module']",
+      "index": 3
+    },
+    {
+      "action": "check-multiple",
+      "selector": "[test-id='select-permission']",
+      "contains": "DEVICE_MANAGEMENT_CRUDL"
+    },
+    {
+      "action": "check",
+      "selector": "[test-id='select-module']",
+      "index": 5
+    },
+    {
+      "action": "check-multiple",
+      "selector": "[test-id='select-permission']",
+      "contains": "DEVICE_BULK_MANAGEMENT_CRUDL"
+    },
+    {
+      "action": "check",
+      "selector": "[test-id='select-module']",
+      "index": 6
+    },
+
+    {
+      "action": "check-multiple",
+      "selector": "[test-id='select-permission']",
+      "contains": "READS_MANAGEMENT_CRUDL"
+    },
+  
+    {
+      "action": "check",
+      "selector": "[test-id='select-module']",
+      "index": 7
+    },
+    {
+      "action": "check-multiple",
+      "selector": "[test-id='select-permission']",
+      "contains": "CERTIFICATE_LOG_MANAGEMENT_CRUDL"
+    },
+    {
+      "action": "check",
+      "selector": "[test-id='select-module']",
+      "index": 10
+    },
+    {
+      "action": "check-multiple",
+      "selector": "[test-id='select-permission']",
+      "contains": "PASSWORD_MANAGEMENT_CRUDL"
+    },
+    {
+      "action": "submit",
+      "selector": "[test-id='submit-permission']"
+    }
+  ]
+
