--- cypress/fixtures/certificate-filter.js
@@ -1,12 +1,0 @@
-[
-  {
-    action: 'click',
-    selector: '[test-id="all-certificates"]',
-  },
-
-  {
-    action: 'select',
-    selector: '[test-id="select-devices"]',
-    option: '[test-id="device-options"]'
-  }
-]
