+++ cypress/fixtures/certificate-filter.json
@@ -1,0 +1,12 @@
+[
+  {
+    "action": "click",
+    "selector": "[test-id='all-certificates']"
+  },
+
+  {
+    "action": "select",
+    "selector": "[test-id='select-devices']",
+    "option": "[test-id='device-options']"
+  }
+]
