--- cypress/fixtures/certificate.js
@@ -1,14 +1,0 @@
-[
-  {
-    action: 'click',
-    selector: '[test-id="reservations"]'
-  },
-  {
-    action: 'click',
-    selector: '[test-id="all-reservations"]'
-  },
-  {
-    action: 'click',
-    selector: '[test-id="certificate"]'
-  }
-]
