+++ cypress/fixtures/certificate.json
@@ -1,0 +1,14 @@
+[
+  {
+    "action": "click",
+    "selector": "[test-id='reservations']"
+  },
+  {
+    "action": "click",
+    "selector": "[test-id='all-reservations']"
+  },
+  {
+    "action": "click",
+    "selector": "[test-id='certificate']"
+  }
+]
