+++ cypress/fixtures/developer-user-login.json
@@ -1,0 +1,20 @@
+[
+  {
+    "action": "type",
+    "selector": "[test-id='login-username']",
+    "value": "developer@energy.org"
+  },
+  {
+    "action": "type",
+    "selector": "[test-id='login-password']",
+    "value": "1234Test@tst"
+  },
+  {
+    "action": "click",
+    "selector": "[test-id='hide-password']"
+  },
+  {
+    "action": "click",
+    "selector": "[test-id='login-submit']"
+  }
+]
