+++ tsconfig.json
@@ -18,6 +18,7 @@
     "moduleResolution": "node",
     "importHelpers": true,
     "target": "ES2022",
+    "types": ["cypress"],
     "module": "ES2022",
     "allowSyntheticDefaultImports": true,
     "useDefineForClassFields": false,
